BasicContentRenderer => ContentRenderer
