+++ app/services/content_renderer.rb
@@ -1,28 +1,45 @@
-# Render markdown with frontmatter (Articles)
+# render markdown for Article's, DisplayAd's, Comment's
 class ContentRenderer
-  class_attribute :fixer, default: MarkdownProcessor::Fixer::FixAll
-  class_attribute :front_matter_parser, default: FrontMatterParser::Parser.new(:md)
+  Result = Struct.new(:front_matter, :reading_time, :processed_html, keyword_init: true)
+
   class_attribute :processor, default: MarkdownProcessor::Parser
+  class_attribute :front_matter_parser, default: FrontMatterParser::Parser.new(:md)
 
   class ContentParsingError < StandardError
   end
 
-  attr_reader :input, :source, :user
-  attr_accessor :reading_time, :front_matter
-
-  def initialize(input, source:, user:)
+  def initialize(input, source:, user: nil, fixer: MarkdownProcessor::Fixer::FixAll)
     @input = input || ""
     @source = source
     @user = user
+    @fixer = fixer
   end
 
-  def process(link_attributes: {}, calculate_reading_time: false)
+  def process(link_attributes: {}, sanitize_options: {},
+              prefix_images_options: { width: 800, synchronous_detail_detection: false })
+    fixed = fixer.call(input)
+    processed = processor.new(fixed, source: source, user: user)
+
+    processed_html = processed.finalize(link_attributes: link_attributes,
+                                        sanitize_options: sanitize_options,
+                                        prefix_images_options: prefix_images_options)
+
+    Result.new(front_matter: nil, processed_html: processed_html, reading_time: 0)
+  rescue StandardError => e
+    raise ContentParsingError, e.message
+  end
+
+  def process_article
     fixed = fixer.call(input)
     parsed = front_matter_parser.call(fixed)
-    self.front_matter = parsed.front_matter
+    front_matter = parsed.front_matter
     processed = processor.new(parsed.content, source: source, user: user)
-    self.reading_time = processed.calculate_reading_time if calculate_reading_time
-    processed.finalize(link_attributes: link_attributes)
+
+    reading_time = processed.calculate_reading_time
+
+    processed_html = processed.finalize
+
+    Result.new(front_matter: front_matter, processed_html: processed_html, reading_time: reading_time)
   rescue StandardError => e
     raise ContentParsingError, e.message
   end
@@ -30,9 +47,13 @@ class ContentRenderer
   def has_front_matter?
     fixed = fixer.call(input)
     parsed = front_matter_parser.call(fixed)
-    self.front_matter = parsed.front_matter
+    front_matter = parsed.front_matter
     front_matter.any? && front_matter["title"].present?
   rescue ContentRenderer::ContentParsingError
     true
   end
+
+  private
+
+  attr_reader :fixer, :input, :user, :source
 end
